@@ -12,7 +12,7 @@
         "intervention/image": "^2.5"
     },
     "require-dev": {
-        "laravel/framework": "^6.0"
+        "laravel/framework": "^8.0"
       },
     "keywords": ["image", "media", "media-explorer", "upload", "resize"],
     "license": "MIT"
